+++ GameOfThrows/GameScene.swift
@@ -4,19 +4,19 @@
 
 import SpriteKit
 
-class GameScene: SKScene, SKPhysicsContactDelegate{
-    let verticalPipeGap = 150.0
-    
-    var bird:SKSpriteNode!
-    var skyColor:SKColor!
-    var pipeTextureUp:SKTexture!
-    var pipeTextureDown:SKTexture!
-    var movePipesAndRemove:SKAction!
-    var moving:SKNode!
-    var pipes:SKNode!
-    var canRestart = Bool()
-    var scoreLabelNode:SKLabelNode!
-    var score = NSInteger()
+class GameScene: SKScene, SKPhysicsContactDelegate {
+    let verticalPipeGap: CGFloat = 150.0
+
+    var bird: SKSpriteNode!
+    var skyColor: SKColor!
+    var pipeTextureUp: SKTexture!
+    var pipeTextureDown: SKTexture!
+    var movePipesAndRemove: SKAction!
+    var moving: SKNode!
+    var pipes: SKNode!
+    var canRestart = false
+    var scoreLabelNode: SKLabelNode!
+    var score = 0
     
     let birdCategory: UInt32 = 1 << 0
     let worldCategory: UInt32 = 1 << 1
@@ -24,136 +24,134 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
     let scoreCategory: UInt32 = 1 << 3
 
     func cancelBirdDeathRotation() {
-        bird?.removeActionForKey("deathRotation")
+        bird?.removeAction(forKey: "deathRotation")
     }
 
     func resetScenePresentation() {
         cancelBirdDeathRotation()
-        self.removeActionForKey("spawnPipes")
-        self.removeActionForKey("flash")
-        self.removeAllChildren()
+        removeAction(forKey: "spawnPipes")
+        removeAction(forKey: "flash")
+        removeAllChildren()
     }
-    
-    override func didMoveToView(view: SKView) {
+
+    override func didMove(to view: SKView) {
         resetScenePresentation()
-        
+
         canRestart = false
-        
+
         // setup physics
-        self.physicsWorld.gravity = CGVectorMake( 0.0, -5.0 )
-        self.physicsWorld.contactDelegate = self
-        
+        physicsWorld.gravity = CGVector(dx: 0.0, dy: -5.0)
+        physicsWorld.contactDelegate = self
+
         // setup background color
         skyColor = SKColor(red: 81.0/255.0, green: 192.0/255.0, blue: 201.0/255.0, alpha: 1.0)
-        self.backgroundColor = skyColor
-        
+        backgroundColor = skyColor
+
         moving = SKNode()
-        self.addChild(moving)
+        addChild(moving)
         pipes = SKNode()
         moving.addChild(pipes)
-        
+
         // ground
         let groundTexture = SKTexture(imageNamed: "land")
-        groundTexture.filteringMode = .Nearest // shorter form for SKTextureFilteringMode.Nearest
-        
-        let moveGroundSprite = SKAction.moveByX(-groundTexture.size().width * 2.0, y: 0, duration: NSTimeInterval(0.02 * groundTexture.size().width * 2.0))
-        let resetGroundSprite = SKAction.moveByX(groundTexture.size().width * 2.0, y: 0, duration: 0.0)
-        let moveGroundSpritesForever = SKAction.repeatActionForever(SKAction.sequence([moveGroundSprite,resetGroundSprite]))
-        
-        for i in 0 ..< Int(2.0 + self.frame.size.width / (groundTexture.size().width * 2.0)){
+        groundTexture.filteringMode = .nearest
+
+        let moveGroundSprite = SKAction.moveBy(x: -groundTexture.size().width * 2.0, y: 0, duration: TimeInterval(0.02 * groundTexture.size().width * 2.0))
+        let resetGroundSprite = SKAction.moveBy(x: groundTexture.size().width * 2.0, y: 0, duration: 0.0)
+        let moveGroundSpritesForever = SKAction.repeatForever(SKAction.sequence([moveGroundSprite, resetGroundSprite]))
+
+        for i in 0..<Int(2.0 + frame.size.width / (groundTexture.size().width * 2.0)) {
             let sprite = SKSpriteNode(texture: groundTexture)
             sprite.setScale(2.0)
-            sprite.position = CGPointMake(CGFloat(i) * sprite.size.width, sprite.size.height / 2.0)
-            sprite.runAction(moveGroundSpritesForever)
+            sprite.position = CGPoint(x: CGFloat(i) * sprite.size.width, y: sprite.size.height / 2.0)
+            sprite.run(moveGroundSpritesForever)
             moving.addChild(sprite)
         }
-        
+
         // skyline
         let skyTexture = SKTexture(imageNamed: "sky")
-        skyTexture.filteringMode = .Nearest
-        
-        let moveSkySprite = SKAction.moveByX(-skyTexture.size().width * 2.0, y: 0, duration: NSTimeInterval(0.1 * skyTexture.size().width * 2.0))
-        let resetSkySprite = SKAction.moveByX(skyTexture.size().width * 2.0, y: 0, duration: 0.0)
-        let moveSkySpritesForever = SKAction.repeatActionForever(SKAction.sequence([moveSkySprite,resetSkySprite]))
-        
-        for i in 0 ..< Int(2.0 + self.frame.size.width / (groundTexture.size().width * 2.0)){
+        skyTexture.filteringMode = .nearest
+
+        let moveSkySprite = SKAction.moveBy(x: -skyTexture.size().width * 2.0, y: 0, duration: TimeInterval(0.1 * skyTexture.size().width * 2.0))
+        let resetSkySprite = SKAction.moveBy(x: skyTexture.size().width * 2.0, y: 0, duration: 0.0)
+        let moveSkySpritesForever = SKAction.repeatForever(SKAction.sequence([moveSkySprite, resetSkySprite]))
+
+        for i in 0..<Int(2.0 + frame.size.width / (groundTexture.size().width * 2.0)) {
             let sprite = SKSpriteNode(texture: skyTexture)
             sprite.setScale(2.0)
             sprite.zPosition = -20
-            sprite.position = CGPointMake(CGFloat(i) * sprite.size.width, sprite.size.height / 2.0 + groundTexture.size().height * 2.0)
-            sprite.runAction(moveSkySpritesForever)
+            sprite.position = CGPoint(x: CGFloat(i) * sprite.size.width, y: sprite.size.height / 2.0 + groundTexture.size().height * 2.0)
+            sprite.run(moveSkySpritesForever)
             moving.addChild(sprite)
         }
-        
+
         // create the pipes textures
         pipeTextureUp = SKTexture(imageNamed: "PipeUp")
-        pipeTextureUp.filteringMode = .Nearest
+        pipeTextureUp.filteringMode = .nearest
         pipeTextureDown = SKTexture(imageNamed: "PipeDown")
-        pipeTextureDown.filteringMode = .Nearest
-        
+        pipeTextureDown.filteringMode = .nearest
+
         // create the pipes movement actions
-        let distanceToMove = CGFloat(self.frame.size.width + 2.0 * pipeTextureUp.size().width)
-        let movePipes = SKAction.moveByX(-distanceToMove, y:0.0, duration:NSTimeInterval(0.01 * distanceToMove))
+        let distanceToMove = frame.size.width + 2.0 * pipeTextureUp.size().width
+        let movePipes = SKAction.moveBy(x: -distanceToMove, y: 0.0, duration: TimeInterval(0.01 * distanceToMove))
         let removePipes = SKAction.removeFromParent()
         movePipesAndRemove = SKAction.sequence([movePipes, removePipes])
-        
+
         // spawn the pipes
-        let spawn = SKAction.runBlock({ [weak self] in self?.spawnPipes() })
-        let delay = SKAction.waitForDuration(NSTimeInterval(2.0))
+        let spawn = SKAction.run { [weak self] in self?.spawnPipes() }
+        let delay = SKAction.wait(forDuration: 2.0)
         let spawnThenDelay = SKAction.sequence([spawn, delay])
-        let spawnThenDelayForever = SKAction.repeatActionForever(spawnThenDelay)
-        self.runAction(spawnThenDelayForever, withKey: "spawnPipes")
-        
+        let spawnThenDelayForever = SKAction.repeatForever(spawnThenDelay)
+        run(spawnThenDelayForever, withKey: "spawnPipes")
+
         // setup our bird
         let birdTexture1 = SKTexture(imageNamed: "bird-01")
-        birdTexture1.filteringMode = .Nearest
+        birdTexture1.filteringMode = .nearest
         let birdTexture2 = SKTexture(imageNamed: "bird-02")
-        birdTexture2.filteringMode = .Nearest
-        
-        let anim = SKAction.animateWithTextures([birdTexture1, birdTexture2], timePerFrame: 0.2)
-        let flap = SKAction.repeatActionForever(anim)
-        
+        birdTexture2.filteringMode = .nearest
+
+        let anim = SKAction.animate(with: [birdTexture1, birdTexture2], timePerFrame: 0.2)
+        let flap = SKAction.repeatForever(anim)
+
         bird = SKSpriteNode(texture: birdTexture1)
         bird.setScale(2.0)
-        bird.position = CGPoint(x: self.frame.size.width * 0.35, y:self.frame.size.height * 0.6)
-        bird.runAction(flap)
-        
-        
+        bird.position = CGPoint(x: frame.size.width * 0.35, y: frame.size.height * 0.6)
+        bird.run(flap)
+
         bird.physicsBody = SKPhysicsBody(circleOfRadius: bird.size.height / 2.0)
-        bird.physicsBody?.dynamic = true
+        bird.physicsBody?.isDynamic = true
         bird.physicsBody?.allowsRotation = false
-        
+
         bird.physicsBody?.categoryBitMask = birdCategory
         bird.physicsBody?.collisionBitMask = worldCategory | pipeCategory
         bird.physicsBody?.contactTestBitMask = worldCategory | pipeCategory
-        
-        self.addChild(bird)
-        
+
+        addChild(bird)
+
         // create the ground
         let ground = SKNode()
-        ground.position = CGPointMake(0, groundTexture.size().height)
-        ground.physicsBody = SKPhysicsBody(rectangleOfSize: CGSizeMake(self.frame.size.width, groundTexture.size().height * 2.0))
-        ground.physicsBody?.dynamic = false
+        ground.position = CGPoint(x: 0, y: groundTexture.size().height)
+        ground.physicsBody = SKPhysicsBody(rectangleOf: CGSize(width: frame.size.width, height: groundTexture.size().height * 2.0))
+        ground.physicsBody?.isDynamic = false
         ground.physicsBody?.categoryBitMask = worldCategory
-        self.addChild(ground)
-        
+        addChild(ground)
+
         // Initialize label and create a label which holds the score
         score = 0
-        scoreLabelNode = SKLabelNode(fontNamed:"MarkerFelt-Wide")
-        scoreLabelNode.position = CGPointMake( CGRectGetMidX( self.frame ), 3 * self.frame.size.height / 4 )
+        scoreLabelNode = SKLabelNode(fontNamed: "MarkerFelt-Wide")
+        scoreLabelNode.position = CGPoint(x: frame.midX, y: 3 * frame.size.height / 4)
         scoreLabelNode.zPosition = 100
         scoreLabelNode.text = String(score)
-        self.addChild(scoreLabelNode)
-        
+        addChild(scoreLabelNode)
     }
 
-    override func willMoveFromView(view: SKView) {
+    override func willMove(from view: SKView) {
         canRestart = false
         moving?.speed = 0
         cancelBirdDeathRotation()
-        self.removeActionForKey("spawnPipes")
-        self.removeActionForKey("flash")
-        self.physicsWorld.contactDelegate = nil
+        removeAction(forKey: "spawnPipes")
+        removeAction(forKey: "flash")
+        physicsWorld.contactDelegate = nil
     }
     
     func spawnPipes() {
@@ -162,44 +160,41 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         }
 
         let pipePair = SKNode()
-        pipePair.position = CGPointMake( self.frame.size.width + pipeTextureUp.size().width * 2, 0 )
+        pipePair.position = CGPoint(x: frame.size.width + pipeTextureUp.size().width * 2, y: 0)
         pipePair.zPosition = -10
-        
-        let height = max(UInt32(UInt(self.frame.size.height / 4)), 1)
+
+        let height = max(UInt32(frame.size.height / 4), 1)
         let y = arc4random_uniform(height) + height
-        
+
         let pipeDown = SKSpriteNode(texture: pipeTextureDown)
         pipeDown.setScale(2.0)
-        pipeDown.position = CGPointMake(0.0, CGFloat(Double(y)) + pipeDown.size.height + CGFloat(verticalPipeGap))
-        
-        
-        pipeDown.physicsBody = SKPhysicsBody(rectangleOfSize: pipeDown.size)
-        pipeDown.physicsBody?.dynamic = false
+        pipeDown.position = CGPoint(x: 0.0, y: CGFloat(y) + pipeDown.size.height + verticalPipeGap)
+
+        pipeDown.physicsBody = SKPhysicsBody(rectangleOf: pipeDown.size)
+        pipeDown.physicsBody?.isDynamic = false
         pipeDown.physicsBody?.categoryBitMask = pipeCategory
         pipeDown.physicsBody?.contactTestBitMask = birdCategory
         pipePair.addChild(pipeDown)
-        
+
         let pipeUp = SKSpriteNode(texture: pipeTextureUp)
         pipeUp.setScale(2.0)
-        pipeUp.position = CGPointMake(0.0, CGFloat(Double(y)))
-        
-        pipeUp.physicsBody = SKPhysicsBody(rectangleOfSize: pipeUp.size)
-        pipeUp.physicsBody?.dynamic = false
+        pipeUp.position = CGPoint(x: 0.0, y: CGFloat(y))
+
+        pipeUp.physicsBody = SKPhysicsBody(rectangleOf: pipeUp.size)
+        pipeUp.physicsBody?.isDynamic = false
         pipeUp.physicsBody?.categoryBitMask = pipeCategory
         pipeUp.physicsBody?.contactTestBitMask = birdCategory
         pipePair.addChild(pipeUp)
-        
+
         let contactNode = SKNode()
-        contactNode.position = CGPointMake( pipeDown.size.width + bird.size.width / 2, CGRectGetMidY( self.frame ) )
-        contactNode.physicsBody = SKPhysicsBody(rectangleOfSize: CGSizeMake( pipeUp.size.width, self.frame.size.height ))
-        contactNode.physicsBody?.dynamic = false
+        contactNode.position = CGPoint(x: pipeDown.size.width + bird.size.width / 2, y: frame.midY)
+        contactNode.physicsBody = SKPhysicsBody(rectangleOf: CGSize(width: pipeUp.size.width, height: frame.size.height))
+        contactNode.physicsBody?.isDynamic = false
         contactNode.physicsBody?.categoryBitMask = scoreCategory
         contactNode.physicsBody?.contactTestBitMask = birdCategory
         pipePair.addChild(contactNode)
-        
-        pipePair.runAction(movePipesAndRemove)
+        pipePair.run(movePipesAndRemove)
         pipes.addChild(pipePair)
-        
     }
 
     func shouldSpawnPipes() -> Bool {
@@ -215,7 +210,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         return moving != nil && moving.speed > 0
     }
     
-    func resetScene (){
+    func resetScene() {
         guard let bird = bird,
             let pipes = pipes,
             let moving = moving,
@@ -223,11 +218,11 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
             return
         }
 
-        bird.removeActionForKey("deathRotation")
+        bird.removeAction(forKey: "deathRotation")
 
         // Move bird to original position and reset velocity
-        bird.position = CGPointMake(self.frame.size.width / 2.5, CGRectGetMidY(self.frame))
-        bird.physicsBody?.velocity = CGVectorMake( 0, 0 )
+        bird.position = CGPoint(x: frame.size.width / 2.5, y: frame.midY)
+        bird.physicsBody?.velocity = CGVector(dx: 0, dy: 0)
         bird.physicsBody?.collisionBitMask = worldCategory | pipeCategory
         bird.speed = 1.0
         bird.zRotation = 0.0
@@ -247,11 +242,11 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         moving.speed = 1
     }
     
-    override func touchesBegan(touches: Set<UITouch>, withEvent event: UIEvent?) {        /* Called when a touch begins */
+    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         if isGameplayRunning() {
             applyBirdImpulse()
-        }else if canRestart {
-            self.resetScene()
+        } else if canRestart {
+            resetScene()
         }
     }
 
@@ -260,26 +255,26 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
             return
         }
 
-        physicsBody.velocity = CGVectorMake(0, 0)
-        physicsBody.applyImpulse(CGVectorMake(0, 30))
+        physicsBody.velocity = CGVector(dx: 0, dy: 0)
+        physicsBody.applyImpulse(CGVector(dx: 0, dy: 30))
     }
-    
+
     // TODO: Move to utilities somewhere. There's no reason this should be a member function
-    func clamp(min: CGFloat, max: CGFloat, value: CGFloat) -> CGFloat {
-        if( value > max ) {
-            return max
-        } else if( value < min ) {
-            return min
+    func clamp(minimum: CGFloat, maximum: CGFloat, value: CGFloat) -> CGFloat {
+        if value > maximum {
+            return maximum
+        } else if value < minimum {
+            return minimum
         } else {
             return value
         }
     }
 
-    func bodyMatchesCategory(body: SKPhysicsBody, category: UInt32) -> Bool {
-        return ( body.categoryBitMask & category ) == category
+    func bodyMatchesCategory(_ body: SKPhysicsBody, category: UInt32) -> Bool {
+        return (body.categoryBitMask & category) == category
     }
 
-    func scoreContactNode(contact: SKPhysicsContact) -> SKNode? {
+    func scoreContactNode(_ contact: SKPhysicsContact) -> SKNode? {
         let bodyAIsScore = bodyMatchesCategory(contact.bodyA, category: scoreCategory)
         let bodyBIsScore = bodyMatchesCategory(contact.bodyB, category: scoreCategory)
         let bodyAIsBird = bodyMatchesCategory(contact.bodyA, category: birdCategory)
@@ -296,7 +291,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         return nil
     }
 
-    func isBirdCollisionContact(contact: SKPhysicsContact) -> Bool {
+    func isBirdCollisionContact(_ contact: SKPhysicsContact) -> Bool {
         let bodyAIsBird = bodyMatchesCategory(contact.bodyA, category: birdCategory)
         let bodyBIsBird = bodyMatchesCategory(contact.bodyB, category: birdCategory)
         let bodyAIsObstacle = bodyMatchesCategory(contact.bodyA, category: worldCategory) ||
@@ -306,9 +301,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
 
         return (bodyAIsBird && bodyBIsObstacle) || (bodyBIsBird && bodyAIsObstacle)
     }
-    
-    
-    override func update(currentTime: CFTimeInterval) {
+    override func update(_ currentTime: TimeInterval) {
         /* Called before each frame is rendered */
         if !isGameplayRunning() {
             return
@@ -319,10 +312,14 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         }
 
         let verticalVelocity = physicsBody.velocity.dy
-        bird.zRotation = self.clamp( -1, max: 0.5, value: verticalVelocity * ( verticalVelocity < 0 ? 0.003 : 0.001 ) )
+        bird.zRotation = clamp(
+            minimum: -1,
+            maximum: 0.5,
+            value: verticalVelocity * (verticalVelocity < 0 ? 0.003 : 0.001)
+        )
     }
-    
-    func didBeginContact(contact: SKPhysicsContact) {
+
+    func didBegin(_ contact: SKPhysicsContact) {
         if !isGameplayRunning() {
             return
         }
@@ -342,7 +339,10 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
             scoreLabelNode.text = String(score)
 
             // Add a little visual feedback for the score increment
-            scoreLabelNode.runAction(SKAction.sequence([SKAction.scaleTo(1.5, duration:NSTimeInterval(0.1)), SKAction.scaleTo(1.0, duration:NSTimeInterval(0.1))]))
+            scoreLabelNode.run(SKAction.sequence([
+                SKAction.scale(to: 1.5, duration: 0.1),
+                SKAction.scale(to: 1.0, duration: 0.1)
+            ]))
             return
         }
 
@@ -353,19 +353,27 @@ class GameScene: SKScene, SKPhysicsContactDelegate{
         moving.speed = 0
 
         bird.physicsBody?.collisionBitMask = worldCategory
-        let deathRotation = SKAction.rotateByAngle(CGFloat(M_PI) * CGFloat(bird.position.y) * 0.01, duration:1)
-        let stopBird = SKAction.runBlock({ bird.speed = 0 })
-        bird.runAction(SKAction.sequence([deathRotation, stopBird]), withKey: "deathRotation")
+        let deathRotation = SKAction.rotate(byAngle: .pi * bird.position.y * 0.01, duration: 1)
+        let stopBird = SKAction.run { bird.speed = 0 }
+        bird.run(SKAction.sequence([deathRotation, stopBird]), withKey: "deathRotation")
 
 
         // Flash background if contact is detected
-        self.removeActionForKey("flash")
-        self.runAction(SKAction.sequence([SKAction.repeatAction(SKAction.sequence([SKAction.runBlock({
-            self.backgroundColor = SKColor(red: 1, green: 0, blue: 0, alpha: 1.0)
-            }),SKAction.waitForDuration(NSTimeInterval(0.05)), SKAction.runBlock({
-                self.backgroundColor = skyColor
-                }), SKAction.waitForDuration(NSTimeInterval(0.05))]), count:4), SKAction.runBlock({
-                    self.canRestart = true
-                    })]), withKey: "flash")
+        removeAction(forKey: "flash")
+        run(SKAction.sequence([
+            SKAction.repeat(SKAction.sequence([
+                SKAction.run {
+                    self.backgroundColor = SKColor(red: 1, green: 0, blue: 0, alpha: 1.0)
+                },
+                SKAction.wait(forDuration: 0.05),
+                SKAction.run {
+                    self.backgroundColor = skyColor
+                },
+                SKAction.wait(forDuration: 0.05)
+            ]), count: 4),
+            SKAction.run {
+                self.canRestart = true
+            }
+        ]), withKey: "flash")
     }
 }