+++ GameOfThrows/GameViewController.swift
@@ -5,48 +5,26 @@
 import UIKit
 import SpriteKit
 
-extension SKNode {
-    class func unarchiveFromFile(file : NSString) -> SKNode? {
-        guard let path = NSBundle.mainBundle().pathForResource(file as String, ofType: "sks"),
-            let sceneData = try? NSData(contentsOfFile: path, options: .DataReadingMappedIfSafe) else {
-            return nil
-        }
-
-        let archiver = NSKeyedUnarchiver(forReadingWithData: sceneData)
-        
-        archiver.setClass(self.classForKeyedUnarchiver(), forClassName: "SKScene")
-        guard let scene = archiver.decodeObjectForKey(NSKeyedArchiveRootObjectKey) as? GameScene else {
-            archiver.finishDecoding()
-            return nil
-        }
-
-        archiver.finishDecoding()
-        return scene
-    }
-}
-
 class GameViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        if let scene = GameScene.unarchiveFromFile("GameScene") as? GameScene {
-            // Configure the view.
-            guard let skView = self.view as? SKView else {
-                return
-            }
-
-            skView.showsFPS = false
-            skView.showsNodeCount = false
-            
-            /* Sprite Kit applies additional optimizations to improve rendering performance */
-            skView.ignoresSiblingOrder = true
-            
-            /* Set the scale mode to scale to fit the window */
-            scene.scaleMode = .AspectFill
-            
-            skView.presentScene(scene)
+        guard let skView = view as? SKView,
+            let scene = GameScene(fileNamed: "GameScene") else {
+            return
         }
+
+        skView.showsFPS = false
+        skView.showsNodeCount = false
+
+        /* Sprite Kit applies additional optimizations to improve rendering performance */
+        skView.ignoresSiblingOrder = true
+
+        /* Set the scale mode to scale to fit the window */
+        scene.scaleMode = .aspectFill
+
+        skView.presentScene(scene)
     }
 
     override func shouldAutorotate() -> Bool {
@@ -54,10 +32,10 @@ class GameViewController: UIViewController {
     }
 
     override func supportedInterfaceOrientations() -> UIInterfaceOrientationMask {
-        if UIDevice.currentDevice().userInterfaceIdiom == .Phone {
-            return UIInterfaceOrientationMask.AllButUpsideDown
+        if UIDevice.current.userInterfaceIdiom == .phone {
+            return .allButUpsideDown
         } else {
-            return UIInterfaceOrientationMask.All
+            return .all
         }
     }